Patching sampling rate in tests
So, @hdykins , it turns out that setTimeout is "_WILDLY_" inaccurate, and there's no that a 10ms counter is ever going to be consistent. ¯\_(ツ)_/¯

We might have to file it away as a bug for now, but as I don't think we've ever needed that kind precision (I think), I upped the number by six.

So. The motivation here is 16ms is ~roughly 1/60s, and it seems to be consistenly performant at that level.